feat: ✨ Extract libplatform
Fix sections alignment in linker scripts for objcopy - each section should have
explicit alignment, otherwise rust-objcopy will not align it properly in binary output.
(see libs/platform/src/platform/raspberrypi/linker/kernel.ld)
