+++ libs/platform/src/platform/raspberrypi/exception/mod.rs
