--- machine/src/platform/raspberrypi/device_driver/bcm/mod.rs
@@ -1,17 +1,0 @@
-// SPDX-License-Identifier: MIT OR Apache-2.0
-//
-// Copyright (c) 2018-2022 Andre Richter <andre.o.richter@gmail.com>
-
-//! BCM driver top level.
-
-pub mod gpio;
-#[cfg(feature = "rpi3")]
-pub mod interrupt_controller;
-// pub mod mailbox;
-pub mod mini_uart;
-pub mod pl011_uart;
-// pub mod power;
-
-#[cfg(feature = "rpi3")]
-pub use interrupt_controller::*;
-pub use {gpio::*, mini_uart::*, pl011_uart::*};
