+++ nucleus/build.rs
@@ -1,6 +1,6 @@
 //! This build script is used to link main kernel binary.
 
-const LINKER_SCRIPT: &str = "machine/src/platform/raspberrypi/linker/kernel.ld";
+const LINKER_SCRIPT: &str = "libs/platform/src/platform/raspberrypi/linker/kernel.ld";
 const LINKER_SCRIPT_AUX: &str = "libs/exception/src/arch/aarch64/linker/aarch64-exceptions.ld";
 
 fn main() {
